Zmiana typu zmiennej dla wieku

+++ Zad1_ProgZaaw_MarekGajda/Models/Users.cs
@@ -20,6 +20,6 @@ namespace Zad1_ProgZaaw_MarekGajda.Models
         [Display(Name = "E-Mail")]
         public string Mail { get; set; }
         [Display(Name = "Wiek")]
-        public decimal Age { get; set; }
+        public int Age { get; set; }
     }
 }
